third_party/WebKit/LayoutTests/fast/dom/shadow/content-element-api-expected.txt

This test ensure that the content element is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS contentSimplest.select is ''
PASS contentSimplest.select is 'foo'
PASS contentSimplest.getAttribute('select') is 'bar'
PASS 0 <= contentSimplest.constructor.toString().indexOf('HTMLContentElement') is true
PASS window.HTMLContentElement is contentSimplest.constructor
PASS successfullyParsed is true

TEST COMPLETE
